sitemap_pro/OCRappBackupFile/251220_Before_Change/requirements.txt
# Core Dependencies
click>=8.0.0
python-dotenv>=0.19.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# PDF Processing
pdf2image>=1.16.0
PyMuPDF>=1.23.0

# Cloud Services
google-cloud-vision>=3.0.0
gspread>=5.10.0
oauth2client>=4.1.3

# Reports
openpyxl>=3.1.0

# Web Scraping
playwright>=1.40.0

# GUI
customtkinter>=5.2.0
